+++ lib/oneview-sdk/exceptions.rb
@@ -12,64 +12,80 @@
 # Contains all the custom Exception classes
 module OneviewSDK
   # Error class allowing storage of a data attribute
-  class OVError < ::StandardError
+  class OneViewError < ::StandardError
     attr_accessor :data
+    MESSAGE = '(No message)'.freeze
 
-    def initialize(msg = nil, data = nil)
+    def initialize(msg = self.class::MESSAGE, data = nil)
       @data = data
       super(msg)
     end
 
     # Shorthand method to raise an error.
     # @example
-    #   OneviewSDK::OVError.raise! 'Message', { data: 'stuff' }
-    def self.raise!(msg = nil, data = nil)
+    #   OneviewSDK::OneViewError.raise! 'Message', { data: 'stuff' }
+    def self.raise!(msg = self::MESSAGE, data = nil)
       raise new(msg, data)
     end
   end
 
-  class ConnectionError < OVError # Cannot connect to client/resource
+  class ConnectionError < OneViewError # Cannot connect to client/resource
+    MESSAGE = 'Cannot connect to client/resource'.freeze
   end
 
-  class InvalidURL < OVError # URL is invalid
+  class InvalidURL < OneViewError # URL is invalid
+    MESSAGE = 'URL is invalid'.freeze
   end
 
-  class InvalidClient < OVError # Client configuration is invalid
+  class InvalidClient < OneViewError # Client configuration is invalid
+    MESSAGE = 'Client configuration is invalid'.freeze
   end
 
-  class InvalidResource < OVError # Failed resource validations
+  class InvalidResource < OneViewError # Failed resource validations
+    MESSAGE = 'Failed resource validations'.freeze
   end
 
-  class IncompleteResource < OVError # Missing required resource data to complete action
+  class IncompleteResource < OneViewError # Missing required resource data to complete action
+    MESSAGE = 'Missing required resource data to complete action'.freeze
   end
 
-  class MethodUnavailable < OVError # Resource does not support this method
+  class MethodUnavailable < OneViewError # Resource does not support this method
+    MESSAGE = 'Resource does not support this method'.freeze
   end
 
-  class UnsupportedVariant < OVError # Variant is not supported
+  class UnsupportedVariant < OneViewError # Variant is not supported
+    MESSAGE = 'Variant is not supported'.freeze
   end
 
-  class UnsupportedVersion < OVError # Resource not supported on this API version
+  class UnsupportedVersion < OneViewError # Resource not supported on this API version
+    MESSAGE = 'Resource not supported on this API version'.freeze
   end
 
-  class InvalidRequest < OVError # Could not make request
+  class InvalidRequest < OneViewError # Could not make request
+    MESSAGE = 'Could not make request'.freeze
   end
 
-  class BadRequest < OVError # 400
+  class BadRequest < OneViewError # 400
+    MESSAGE = '400'.freeze
   end
 
-  class Unauthorized < OVError # 401
+  class Unauthorized < OneViewError # 401
+    MESSAGE = '401'.freeze
   end
 
-  class NotFound < OVError # 404
+  class NotFound < OneViewError # 404
+    MESSAGE = '404'.freeze
   end
 
-  class RequestError < OVError # Other bad response codes
+  class RequestError < OneViewError # Other bad response codes
+    MESSAGE = 'Bad response code'.freeze
   end
 
-  class TaskError < OVError # Task ended in a bad state
+  class TaskError < OneViewError # Task ended in a bad state
+    MESSAGE = 'Task ended in a bad state'.freeze
   end
 
-  class InvalidFormat < OVError # File format is invalid
+  class InvalidFormat < OneViewError # File format is invalid
+    MESSAGE = 'File format is invalid'.freeze
   end
 end